refactor code
